index entity DataProvider update and support to use proxyHelper
